error on virtual directory that does not exist
On the first call to find(), index the directories in the box.
When returning a virtual directory, first check to see if that
directory exists, if not return a not found error.

This will cause http.FileServer to return a 404 on a
directory that does not exist in the box.