Reject, early on, a directory that can't have any Go files accepted as prospects.
This also warns the user for such cases (several times, it seems; not
sure why the dir-walking code is invoked more than once).

Note that a workaround for having e.g. `.go/` as a component in the
path to Go source tree being "packed" is, for example:

```
$ packr2 -- .
$ go build
```

Even though there will still be warnings, `.` as an explicit directory
name will (as confirmed by the `-v` option) pick up `.go` files as
legitimate prospects, because the full path is not included in the
prospect check(s).