Check for `go.mod` when searching for files.
It appears that `GoFiles` from
[`Package`](https://godoc.org/go/build#Package) is _NOT_ module aware.
Therefore the check in `v2/jam/parser/finder.go:87` would always return
and not find files. The only way in which I think we can reliably check
if a project is using modules is to check for the presence of `go.mod`
in the package root. Thus, I have added the guard to check for `go.mod`
before we rely on `pkg.GoFiles`.