Always accept initial path
Always accept the path that packr2 should work on, even if it contains
ignored path segments.
Ignore prefixes continue to work as before, but only for sub paths.
This enables the use of packr2 in CI builds where the project is placed
in a temporary location for building, which contains '.' or '_' prefix.